flask3/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
alembic==1.15.2
    # via flask-migrate
blinker==1.9.0
    # via flask
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
flask==3.1.0
    # via
    #   flask2 (pyproject.toml)
    #   flask-migrate
    #   flask-sqlalchemy
flask-migrate==4.1.0
    # via flask2 (pyproject.toml)
flask-sqlalchemy==3.1.1
    # via
    #   flask2 (pyproject.toml)
    #   flask-migrate
greenlet==3.1.1
    # via sqlalchemy
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
mako==1.3.9
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.26.1
    # via flask2 (pyproject.toml)
packaging==24.2
    # via marshmallow
psycopg2-binary==2.9.10
    # via flask2 (pyproject.toml)
python-dotenv==1.1.0
    # via flask2 (pyproject.toml)
sqlalchemy==2.0.40
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.13.1
    # via
    #   alembic
    #   sqlalchemy
werkzeug==3.1.3
    # via flask
